app/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
pydantic==2.9.2
mlflow==2.17.2
PyYAML==6.0.2
joblib==1.4.2
imbalanced-learn==0.12.4
prometheus-client==0.21.0
